Warnings.

+++ src/hashes/rmd128.c
@@ -382,7 +382,7 @@ int rmd128_test(void)
 
    for (i = 0; i < (int)LTC_ARRAY_SIZE(tests); i++) {
        rmd128_init(&md);
-       rmd128_process(&md, (unsigned char *)tests[i].msg, XSTRLEN(tests[i].msg));
+       rmd128_process(&md, (unsigned char *)tests[i].msg, (unsigned long)XSTRLEN(tests[i].msg));
        rmd128_done(&md, tmp);
        if (ltc_compare_testvector(tmp, sizeof(tmp), tests[i].hash, sizeof(tests[i].hash), "RIPEMD128", i)) {
           return CRYPT_FAIL_TESTVECTOR;
